ai-agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
